# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
        "/product-categories": {
            "get": {
                "tags": [
                    "product-categories"
                ],
                "summary": "Get all product categories",
                "description": "This can only be done by the logged in user.",
                "operationId": "getProductCategories",
                "parameters": [
                    {
                        "name": "includeDeleted",
                        "in": "query",
                        "description": "Should include deleted product category",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "boolean",
                            "default": "false"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductCategory"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductCategory"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "product-categories"
                ],
                "summary": "create product",
                "description": "This can only be done by the logged in user.",
                "operationId": "createProductCategory",
                "requestBody": {
                    "description": "Create product category object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCategory"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get all products",
                "description": "This can only be done by the logged in user.",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "includeDeleted",
                        "in": "query",
                        "description": "Should include deleted product",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "boolean",
                            "default": "false"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "create product",
                "description": "This can only be done by the logged in user.",
                "operationId": "createProduct",
                "requestBody": {
                    "description": "Create product object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/products/{productId}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Find product by ID",
                "description": "Returns a single product",
                "operationId": "showProduct",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of product to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "products"
                ],
                "summary": "update product",
                "description": "This can only be done by the logged in user.",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of product to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update product object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "show user",
                "description": "This can only be done by the logged in user.",
                "operationId": "showUser",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product model",
                "description": "Product model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Product name",
                        "type": "string",
                        "format": "int64"
                    },
                    "category": {
                        "title": "Category",
                        "description": "Product category",
                        "type": "string",
                        "format": "int64"
                    },
                    "quantity": {
                        "title": "Quantity",
                        "description": "Product quantity",
                        "type": "integer",
                        "format": "int64"
                    },
                    "deleted": {
                        "title": "Deleted",
                        "description": "Is product deleted?",
                        "type": "boolean",
                        "format": "int64",
                        "default": "false"
                    },
                    "price": {
                        "title": "Price",
                        "description": "Product price per piece",
                        "type": "number",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
            "ProductCategory": {
                "title": "ProductCategory model",
                "description": "ProductCategory model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Product Category name",
                        "type": "string",
                        "format": "int64"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Product Category description",
                        "type": "string",
                        "format": "int64"
                    },
                    "deleted": {
                        "title": "Deleted",
                        "description": "Is product category deleted?",
                        "type": "boolean",
                        "format": "int64",
                        "default": "false"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User model",
                "description": "User model"
            }
        }
    },
    "tags": [
        {
            "name": "product-categories",
            "description": "product-categories"
        },
        {
            "name": "products",
            "description": "products"
        },
        {
            "name": "user",
            "description": "user"
        }
    ]
}